moved drawer to standalone component
